Fix issue with instancing config from subclasses of HttpClientConfig
