+++ client/src/main/java/se/fortnox/reactivewizard/client/HttpClientConfig.java
@@ -51,7 +51,7 @@ public class HttpClientConfig {
     @JsonProperty("validateCertificates")
     private boolean isValidateCertificates = true;
 
-    private BasicAuth basicAuth;
+    private BasicAuthConfig basicAuth;
 
     public HttpClientConfig() {
     }
@@ -176,40 +176,18 @@ public class HttpClientConfig {
         isValidateCertificates = value;
     }
 
-    public BasicAuth getBasicAuth() {
+    public BasicAuthConfig getBasicAuth() {
         return basicAuth;
     }
 
-    public void setBasicAuth(BasicAuth basicAuth) {
+    public void setBasicAuth(BasicAuthConfig basicAuth) {
         this.basicAuth = basicAuth;
     }
 
     public void setBasicAuth(String username, String password) {
-        this.basicAuth = new BasicAuth()
+        this.basicAuth = new BasicAuthConfig()
             .setUsername(username)
             .setPassword(password);
     }
 
-    public class BasicAuth {
-        private String username;
-        private String password;
-
-        public String getUsername() {
-            return username;
-        }
-
-        public BasicAuth setUsername(String username) {
-            this.username = username;
-            return this;
-        }
-
-        public String getPassword() {
-            return password;
-        }
-
-        public BasicAuth setPassword(String password) {
-            this.password = password;
-            return this;
-        }
-    }
 }
